refactor : Users가 카드를 주는 로직 제거
