+++ src/main/java/blackjack/domain/dto/GameResultDto.java
@@ -1,13 +1,5 @@
 package blackjack.domain.dto;
 
-import blackjack.domain.BlackJack;
-import blackjack.domain.Result;
-import blackjack.domain.Users;
-import blackjack.domain.card.GamePoint;
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.Name;
-import blackjack.domain.user.User;
-
 import java.util.*;
 
 public class GameResultDto {
