+++ src/main/java/blackjack/domain/dto/InitialStatusDto.java
@@ -1,16 +1,6 @@
 package blackjack.domain.dto;
 
-import blackjack.domain.BlackJack;
-import blackjack.domain.Users;
-import blackjack.domain.card.Card;
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.Name;
-import blackjack.domain.user.User;
-
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class InitialStatusDto {
     private final CardDto dealerCard;
